# Time optimization for multimodel module case overlap