ok guys, in my trials and tribulations with dask and iris2 I just found me a speedup of order factor 15 for multimodel if you patch it this way. I tested it and results are perfectly identical to what we'd get before, only much, much FASTER :grin: