@valeriupredoi 2019-02-01
so I thought hmmm maybe this optimization may not scale with data volumes and went to Jasmin and tested the before/after patch with some 8 datasets and lots of years' woth of data and lo behold the results:

```
Before patch
------------
2019-02-01 13:39:56,305 UTC [16179] DEBUG   esmvaltool.preprocessor:197 Running preprocessor step multi_model_statistics
2019-02-01 13:41:04,105 UTC [16179] DEBUG   esmvaltool.preprocessor:197 Running preprocessor step cmor_check_data
dt = 68s
max mem: 14.2GB

After patch
-----------
2019-02-01 13:53:20,080 UTC [22805] DEBUG   esmvaltool.preprocessor:197 Running preprocessor step multi_model_statistics
2019-02-01 13:53:29,316 UTC [22805] DEBUG   esmvaltool.preprocessor:197 Running preprocessor step cmor_check_data
dt = 9s
max mem: 14.1GB 

data comparison
---------------
In [5]: np.mean(c1.data)
Out[5]: 236.8959

In [6]: np.mean(c2.data)
Out[6]: 236.8959
```
I mean, the sci node on Jasmin is a darn powerful one but 68s -> 9s :grin: