@valeriupredoi 2019-02-01
and the memory estimation I came up with in the PR is absolutely bang-on:

R ~ 2.1, F ~ 1.4GB, N = 8
```
In [5]: (2*2.1 + 8)*1.4 - 2*1.4
Out[5]: 14.28
```