@valeriupredoi 2019-02-01
nops, iris1.13, iris2 sees the same improvement in time with the multimodel patch but it's offset by 20-40% slower overall compared to iris1.13 (note that there's no dasks in it yet)