@valeriupredoi 2019-02-01
memory stays the same though, no significant increase in memory from iris1 to iris2 so yay :grin: