@mattiarighi 2019-02-07
I haven't noticed any significant improvement for perfmetrics, but given the tests by @valeriupredoi I would merge this anyway.